requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.2
